EIEIO added and corrected
Working on eieio import

Some parts appear to need refreshing, and a proper suite of automated tests needs building. We also need to address the fact that TLS negotiation looks like it can now be straightforwardly imported from the standard java libraries. In the limit this can be done by using the non-nio libraries because running processes are no longer (post Virtual Threads) a scarce-ish resource.

Starting on ssl experiments