Incrementally restructuring the simple ox.net.ssl module
The plan is to unbundle codecs and encodings, as well as Logging.
My first attempt at promoting Log and Logging to a separate module was not successful, the compiler(s) insisted that there was no such package as ox.logging when compiling eieio; but allowed the same package when compiling ssl and PseudoChannel.